requirements.txt
streamli
pandas
numpy
scikit-learn